+++ Libraries/RANSUROTTO.BLOG.Services/Logging/CustomerActivityService.cs
@@ -118,7 +118,7 @@ namespace RANSUROTTO.BLOG.Services.Logging
         /// </summary>
         /// <param name="activityLogTypeId">活动日志类型标识符</param>
         /// <returns>活动日志类型</returns>
-        public virtual ActivityLogType GetActivityTypeById(long activityLogTypeId)
+        public virtual ActivityLogType GetActivityTypeById(int activityLogTypeId)
         {
             if (activityLogTypeId == 0)
                 return null;
@@ -196,7 +196,7 @@ namespace RANSUROTTO.BLOG.Services.Logging
         /// <param name="ipAddress">IP地址; NULL为不限制</param>
         /// <returns>活动日志列表</returns>
         public virtual IPagedList<ActivityLog> GetAllActivities(DateTime? createdOnFrom = null,
-            DateTime? createdOnTo = null, long? customerId = null, long activityLogTypeId = 0,
+            DateTime? createdOnTo = null, int? customerId = null, int activityLogTypeId = 0,
             int pageIndex = 0, int pageSize = int.MaxValue, string ipAddress = null)
         {
             var query = _activityLogRepository.Table;
@@ -222,7 +222,7 @@ namespace RANSUROTTO.BLOG.Services.Logging
         /// </summary>
         /// <param name="activityLogId">活动日志标识符</param>
         /// <returns>活动日志</returns>
-        public virtual ActivityLog GetActivityById(long activityLogId)
+        public virtual ActivityLog GetActivityById(int activityLogId)
         {
             if (activityLogId == 0)
                 return null;
@@ -273,7 +273,7 @@ namespace RANSUROTTO.BLOG.Services.Logging
         [Serializable]
         public class ActivityLogTypeForCaching
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public string SystemKeyword { get; set; }
             public string Name { get; set; }
             public bool Enabled { get; set; }
